requirements.txt
webrtcvad==2.0.10 
pandas==1.1.3 
ffmpeg-python==0.2.0
youtube_dl==2020.11.12
Flask==1.1.2
flask-wtf
flask-bootstrap
